+++ crates/sncast/src/response/get/mod.rs
@@ -3,6 +3,7 @@ pub mod block;
 pub mod class_hash_at;
 pub mod nonce;
 pub mod spec_version;
+pub mod state_update;
 pub mod transaction;
 pub mod tx_receipt;
 pub mod tx_status;