+++ crates/sncast/tests/e2e/get/mod.rs
@@ -3,6 +3,7 @@ mod block;
 mod class_hash_at;
 mod nonce;
 mod spec_version;
+mod state_update;
 mod transaction;
 mod tx_receipt;
 mod tx_status;